# Conflicts:
#	docker-compose.yml
#	pom.xml
#	src/main/java/com/example/config_service_api/ConfigServiceApiApplication.java
#	src/main/java/com/example/config_service_api/controller/EnvironmentController.java
#	src/main/java/com/example/config_service_api/dto/PageableDto.java
#	src/main/java/com/example/config_service_api/entity/DataEntity.java
#	src/main/java/com/example/config_service_api/entity/EnvironmentEntity.java
#	src/main/java/com/example/config_service_api/repository/DataRepository.java
#	src/main/java/com/example/config_service_api/service/EnvironmentService.java
#	src/main/java/com/example/config_service_api/service/NamespaceService.java
#	src/main/resources/application.properties

+++ src/main/java/com/example/config_service_api/service/NamespaceService.java
@@ -79,7 +79,7 @@ public class NamespaceService {
                 .content(contentDto)
                 .currentPage(namespacesPage.getNumber())
                 .pageSize(namespacesPage.getSize())
-                .totalElements((int) namespacesPage.getTotalElements())
+                .totalElements(namespacesPage.getTotalElements())
                 .totalPages(namespacesPage.getTotalPages())
                 .first(namespacesPage.isFirst())
                 .build();
@@ -210,7 +210,7 @@ public class NamespaceService {
                 .content(contentDto)
                 .currentPage(environmentsPage.getNumber())
                 .pageSize(environmentsPage.getSize())
-                .totalElements((int) environmentsPage.getTotalElements())
+                .totalElements(environmentsPage.getTotalElements())
                 .totalPages(environmentsPage.getTotalPages())
                 .first(environmentsPage.isFirst())
                 .build();
